Touch up build script to add a cleanup task.

+++ build.sh
@@ -6,8 +6,11 @@ YELLOW="\033[1;33m"
 NOCOLOR="\033[0m"
 currentdir=${PWD##*/}
 
-if [ ! -d ./srf ]; then
-	echo "The /srf directory was not found. Creating directory..."
+if [ -d ./srf ]; then
+	echo "Removing old build dir..."
+	rm -r ./srf
+else
+	echo "Creating a new build directory..."
 	mkdir srf
 fi
 
@@ -17,12 +20,16 @@ if [ ! -f ./build-exclude.txt ]; then
 fi
 
 echo "${GREEN}Syncing srf files...${NOCOLOR}"
-rm -r ./srf/*
-rm ./srf-theme.zip
+if [ -f ./srf-theme.zip ]; then
+	rm ./srf-theme.zip
+fi
 rsync -avzu --delete --progress -h --exclude-from="build-exclude.txt" ./* ./srf
 
 echo "${GREEN}Syncing complete.${NOCOLOR}"
 echo "${GREEN}Building zip file...${NOCOLOR}"
 zip -r ./srf-theme.zip ./srf
+
+echo "${GREEN}Cleaning up...${NOCOLOR}"
+rm -r ./srf
 
 echo "${GREEN}Complete! Upload the ./srf-theme.zip file to your Themes in wp-admin.${NOCOLOR}"
